+++ src/RC-Car/source/network_driver.h
@@ -37,15 +37,67 @@
 * so agrees to indemnify Cypress against all liability.
 *******************************************************************************/
 
-#ifndef UDP_SERVER_H_
-#define UDP_SERVER_H_
+#ifndef NETWORK_DRIVER_H
+#define NETWORK_DRIVER_H
 
+#include <inttypes.h>
+#include "utils.h"
+
+
+/*******************************************************************************
+* Macros
+********************************************************************************/
+/* Wi-Fi Credentials: Modify WIFI_SSID, WIFI_PASSWORD, and WIFI_SECURITY_TYPE
+ * to match your Wi-Fi network credentials.
+ * Note: Maximum length of the Wi-Fi SSID and password is set to
+ * CY_WCM_MAX_SSID_LEN and CY_WCM_MAX_PASSPHRASE_LEN as defined in cy_wcm.h file.
+ */
+
+#define WIFI_SSID                                 "WIFI_SSID"
+#define WIFI_PASSWORD                             "WIFI_PASSWORD"
+
+/* Security type of the Wi-Fi access point. See 'cy_wcm_security_t' structure
+ * in "cy_wcm.h" for more details.
+ */
+#define WIFI_SECURITY_TYPE                        CY_WCM_SECURITY_WPA2_AES_PSK
+
+/* Maximum number of connection retries to a Wi-Fi network. */
+#define MAX_WIFI_CONN_RETRIES                     (10u)
+
+/* Wi-Fi re-connection time interval in milliseconds */
+#define WIFI_CONN_RETRY_INTERVAL_MSEC             (1000)
+
+/* UDP server related macros. */
+#define UDP_SERVER_PORT                           (57345)
+#define UDP_SERVER_MAX_PENDING_CONNECTIONS        (3)
+#define UDP_SERVER_RECV_TIMEOUT_MS                (500)
+
+
+/*******************************************************************************
+Protocol definitions 
+********************************************************************************/
+typedef struct __attribute__((__packed__))
+{
+    uint32_t   sequence_id;
+    uint16_t   msg_length;
+
+    struct __attribute__((__packed__))
+    {
+        uint8_t    command;
+        val_type_t data;
+    } payload;
+    
+    uint32_t   crc_32;
+} client_req_t;
+
+
+typedef int (*message_reception_callback)( client_req_t* );
 
 /*******************************************************************************
 * Function Prototypes
 ********************************************************************************/
-int rc_car_init( void );
-void rc_car_app_task(void *arg);
+void set_network_callback(message_reception_callback _callback);
+void network_task(void *arg);
 
 #endif /* UDP_SERVER_H_ */
